webapp/backend/requirements.txt
# Core FastAPI for Cloudflare Workers
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1

# Security & Authentication  
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# HTTP & API Clients
httpx==0.25.2
requests==2.31.0

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Environment Configuration
python-dotenv==1.0.0

# System Monitoring
psutil==5.9.6

# Data Processing
pandas==2.2.2